requirements.txt
Pillow>=2.9.0
pytz>=2015.7
django-ckeditor
django-mptt-graph
django-mptt